This commit supports multiple Wire objects on those boards which have multiple I2C busses. Example: T3.1/T3.2/T-LC - Support 2 so there is Wire and now Wire1 T3.5: has three I2C Busses, so it now has Wire, Wire1 and Wire2 T3.6: has four..... Wire3
For this, I Created a base class for the Wire objects, that has most of the functionality.
There are some virtual functions, which take care of the buss specific
things like, begin/end/setSCL/setSDA.

Also object keeps pointer to the I2C register structure for the buss and
got rid of most statitics.

With my first attempt, I found that this was eating up a significant amount of memory, as each Wire object has something like 80 data bytes associated with it. So I created an internal data structure, where I moved
the majority of the data members to.  This data structure is only allocated (malloc) when you call some members such as begin.  So there is not a large data hit for those programs that include Wire.h and only use 1 buss...

Tested on T3.6 with Wire talking to Wire3 both as Server and as client.  Also
tried Setting Wire to alternate pins to again talk to Wire3 with alternate pins for Wire.

Also tested with a modified version of SSD1306 I2C display and tried on all 4 busses of T3.6, also with Wire1 on T3.2 and T-LC.  Also tried Wire on T-LC.

Resolve issue where was timing out on some calls

Ran into issue that I would try to do a request from the client before
the previous ones stop bit was processed and we were clearing some
status.

Current fix is to loop up to 255 times checking for busy before
continuing...